+++ packages/blade/src/components/Charts/CommonChartComponents/types.ts
@@ -1,0 +1,66 @@
+import type {
+  XAxisProps as RechartsXAxisProps,
+  YAxisProps as RechartsYAxisProps,
+  CartesianGridProps as RechartsCartesianGridProps,
+  Tooltip as RechartsTooltip,
+  Legend as RechartsLegend,
+  ReferenceLineProps as RechartsReferenceLineProps,
+} from 'recharts';
+import type { ComponentProps } from 'react';
+import type { ChartColorCategories, ChartCategoricalEmphasis } from '~tokens/theme/theme';
+
+type ChartReferenceLineProps = {
+  /**
+   * The y-coordinate of the reference line.
+   */
+  y?: RechartsReferenceLineProps['y'];
+  /**
+   * The x-coordinate  of the reference line.
+   */
+  x?: RechartsReferenceLineProps['x'];
+  /**
+   * The label of the reference line.
+   */
+  label: string;
+};
+
+type ChartXAxisProps = Omit<RechartsXAxisProps, 'tick' | 'label' | 'dataKey' | 'stroke'> & {
+  /**
+   * The label of the x-axis.
+   */
+  label?: string;
+  /**
+   * The data key of the x-axis.
+   */
+  dataKey?: string;
+};
+type ChartYAxisProps = Omit<RechartsYAxisProps, 'tick' | 'label' | 'dataKey' | 'stroke'> & {
+  /**
+   * The label of the y-axis.
+   */
+  label?: string;
+  /**
+   * The data key of the y-axis.
+   */
+  dataKey?: string;
+};
+
+type ChartTooltipProps = ComponentProps<typeof RechartsTooltip>;
+type ChartLegendProps = ComponentProps<typeof RechartsLegend>;
+
+type ChartCartesianGridProps = Omit<
+  RechartsCartesianGridProps,
+  'strokeDasharray' | 'verticalFill' | 'horizontalFill'
+>;
+
+type ChartsCategoricalColorToken = `chart.background.categorical.${ChartColorCategories}.${keyof ChartCategoricalEmphasis}`;
+
+export type {
+  ChartReferenceLineProps,
+  ChartXAxisProps,
+  ChartYAxisProps,
+  ChartTooltipProps,
+  ChartLegendProps,
+  ChartCartesianGridProps,
+  ChartsCategoricalColorToken,
+};