+++ packages/blade/src/components/DatePicker/BaseDatePicker.web.tsx
@@ -13,13 +13,14 @@ import { Calendar } from './Calendar.web';
 import { CalendarFooter } from './CalendarFooter.web';
 import { DatePickerInput } from './DateInput.web';
 import { DatePickerFilterChip } from './FilterChipDatePicker/DatePickerFilterChip.web';
-import { PresetDropdown } from './QuickSelection/PresetDropdown.web';
-import { PresetProvider } from './QuickSelection/PresetContext';
+import { renderPresetDropdown } from './QuickSelection/renderPresetDropdown.web';
 import { PresetSideBar } from './QuickSelection/PresetSideBar.web';
+import { usePresetState } from './QuickSelection/usePresetState';
 import { shiftTimezone } from './shiftTimezone';
 import { useDatesState } from './useDatesState';
 import { usePopup } from './usePopup';
 import { convertIntlToDayjsLocale, loadScript } from './utils';
+import { DatePickerProvider } from './DatePickerContext';
 import BaseBox from '~components/Box/BaseBox';
 import { useControllableState } from '~utils/useControllable';
 import { useTheme } from '~utils';
@@ -46,7 +47,8 @@ import { useFilterChipGroupContext } from '~components/Dropdown/FilterChipGroupC
 // Calendar dimensions for consistent layout
 const CALENDAR_HEIGHTS = {
   // Height includes: Calendar grid (6 weeks * ~44px) + header (~48px) + footer actions (~64px) + padding
-  DAY_PICKER_WITH_FOOTER: '447px',
+  // moved to auto since we are exposing footer slot
+  DAY_PICKER_WITH_FOOTER: 'auto',
 } as const;
 
 const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
@@ -84,6 +86,7 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
   labelSuffix,
   labelTrailing,
   showFooterActions = true,
+  footer,
   ...props
 }: DatePickerProps<Type> &
   StyledPropsBlade &
@@ -177,6 +180,13 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
   });
 
   const currentDate = shiftTimezone('add', new Date());
+
+  // Use the hook to get the calculated preset values
+  const { presetStates, selectedPresetLabel, effectiveSelectionType } = usePresetState({
+    presets: presets || [],
+    selectedPreset,
+    currentDate,
+  });
   const hasBothDatesSelected = controlledValue?.[0] && controlledValue?.[1];
   const { listViewSelectedFilters, setListViewSelectedFilters } = useListViewFilterContext();
   const {
@@ -312,11 +322,12 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
     <>
       {shouldRenderPresets ? (
         <PresetSideBar
-          onSelection={(preset) => {
+          onSelection={(preset: (date: Date) => DatesRangeValue) => {
             const presetValue = preset?.(currentDate);
             setControlledValue(presetValue);
             setSelectedPreset(presetValue);
           }}
+          presetStates={presetStates}
         />
       ) : null}
       <BaseBox
@@ -383,6 +394,8 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
               isButtonDisabled={applyButtonDisabled}
               onApply={handleApply}
               onCancel={handleCancel}
+              footer={footer}
+              selectionType={_selectionType}
             />
           ))}
       </BaseBox>
@@ -415,7 +428,7 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
   return (
     <MantineProvider>
       <DatesProvider settings={dateProviderValue}>
-        <PresetProvider presets={presets} selectedPreset={selectedPreset} currentDate={currentDate}>
+        <DatePickerProvider isDatePickerBodyOpen={controllableIsOpen}>
           <BaseBox
             width={inputElementType === 'chip' ? 'fit-content' : '100%'}
             {...getStyledProps(props)}
@@ -475,19 +488,23 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
                 excludeDate={props.excludeDate}
                 minDate={props.minDate}
                 maxDate={props.maxDate}
+                // Effective Selection type should only be use for selectionType 'range'
+                effectiveSelectionType={isSingle ? selectionType : effectiveSelectionType}
                 leadingDropdown={
-                  presets && shouldRenderPresets ? (
-                    <PresetDropdown
-                      onSelection={(preset) => {
-                        const presetValue = preset?.(currentDate);
-                        setControlledValue(presetValue);
-                        setSelectedPreset(presetValue);
-                      }}
-                      onOpenCalendar={() => {
-                        controllableSetIsOpen(() => true);
-                      }}
-                    />
-                  ) : undefined
+                  presets && !isSingle
+                    ? renderPresetDropdown({
+                        onSelection: (preset: (date: Date) => DatesRangeValue) => {
+                          const presetValue = preset?.(currentDate);
+                          setControlledValue(presetValue);
+                          setSelectedPreset(presetValue);
+                        },
+                        onOpenCalendar: () => {
+                          controllableSetIsOpen(() => true);
+                        },
+                        presetStates,
+                        selectedPresetLabel,
+                      })
+                    : undefined
                 }
                 {...makeAnalyticsAttribute(props)}
               />
@@ -503,10 +520,11 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
                 <BottomSheetHeader title={isSingle ? 'Select Date' : 'Select Date Range'} />
                 <BottomSheetBody>
                   {content}
-                  {!isSingle && (
+                  {!isSingle && presets && (
                     <PresetSideBar
                       isMobile
-                      onSelection={(preset) => {
+                      presetStates={presetStates}
+                      onSelection={(preset: (date: Date) => DatesRangeValue) => {
                         const presetValue = preset?.(currentDate);
                         setControlledValue(presetValue);
                         setSelectedPreset(presetValue);
@@ -516,7 +534,12 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
                 </BottomSheetBody>
                 {showFooterActions && (
                   <BottomSheetFooter>
-                    <CalendarFooter onCancel={handleCancel} onApply={handleApply} />
+                    <CalendarFooter
+                      onCancel={handleCancel}
+                      onApply={handleApply}
+                      footer={footer}
+                      selectionType={_selectionType}
+                    />
                   </BottomSheetFooter>
                 )}
               </BottomSheet>
@@ -555,7 +578,7 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
               )
             )}
           </BaseBox>
-        </PresetProvider>
+        </DatePickerProvider>
       </DatesProvider>
     </MantineProvider>
   );